misc/HtmlTestSuite/results/test-0749.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Logical Properties: {max-,min-}inline-size</title>
<link href="mailto:openxu@gmail.com" rel="author" title="Xu Xing"/>
<link href="https://drafts.csswg.org/css-logical-props-1/#logical-dimension-properties" rel="help"/>
<link href="https://drafts.csswg.org/css-writing-modes-3/#logical-to-physical" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

<style>
.block {
  border: 1px solid #000;
}
#block1 {
  inline-size: 40px;
  min-inline-size: 50px;
  max-inline-size: 100px;
}
#block2 {
  inline-size: 100px;
  min-inline-size: 50px;
  max-inline-size: 100px;
}
#block3 {
  inline-size: 120px;
  min-inline-size: 50px;
  max-inline-size: 100px;
}

.override {
  border: 1px solid #000;
}
#override1 {
  inline-size: 100px;
  width: 50px;
}
#override2 {
  width: 50px;
  inline-size: 100px;
}

.table {
  border: 1px solid #000;
  display: table;
}
.tablecell {
  display: table-cell;
}
#table1_cell {
  inline-size: 40px;
  min-inline-size: 50px;
  max-inline-size: 100px;
  block-size: 100px;
  background-color: red;
}
#table2_cell {
  inline-size: 100px;
  min-inline-size: 50px;
  max-inline-size: 100px;
  block-size: 100px;
  background-color: blue;
}
#table3_cell {
  inline-size: 120px;
  min-inline-size: 50px;
  max-inline-size: 100px;
  block-size: 100px;
  background-color: green;
}
</style>

</head><body><div id="log"/>

<h3>Maximum and minimim inline sizes in blocks</h3>
<div>
  <p><code>inline-size</code> &lt; <code>min-inline-size</code></p>
  <div class="block" data-expected-client-height="0" data-expected-client-width="50" id="block1"/>

  <p><code>min-inline-size</code> &lt; <code>inline-size</code> ≤ <code>max-inline-size</code></p>
  <div class="block" data-expected-client-height="0" data-expected-client-width="100" id="block2"/>

  <p><code>inline-size</code> &gt; <code>max-inline-size</code></p>
  <div class="block" data-expected-client-height="0" data-expected-client-width="100" id="block3"/>
</div>

<h3>Overridance of <code>width</code> and <code>inline-size</code></h3>
<div>
  <p>Check that <code>width</code> overrides <code>inline-size</code></p>
  <div class="override" data-expected-client-height="0" data-expected-client-width="50" id="override1"/>

  <p>Check that <code>inline-size</code> overrides <code>width</code></p>
  <div class="override" data-expected-client-height="0" data-expected-client-width="100" id="override2"/>
</div>

<h3>Maximum and minimim inline sizes in table cells</h3>
<div>
  <p><code>inline-size</code> &lt; <code>min-inline-size</code></p>
  <div class="table">
    <div class="tablecell" data-expected-client-height="100" data-expected-client-width="50" id="table1_cell"/>
  </div>

  <p><code>min-inline-size</code> &lt; <code>inline-size</code> ≤ <code>max-inline-size</code></p>
  <div class="table">
    <div class="tablecell" data-expected-client-height="100" data-expected-client-width="100" id="table2_cell"/>
  </div>

  <p><code>inline-size</code> &gt; <code>max-inline-size</code></p>
  <div class="table">
    <div class="tablecell" data-expected-client-height="100" data-expected-client-width="100" id="table3_cell"/>
  </div>
</div>

<script>
checkLayout(".block", false);
checkLayout(".override", false);
checkLayout(".tablecell", false);
done();
</script>
</body></html>
